Remove unused route names and refactor job-related screens to streamline the codebase
