--- lib/features/auth/presentation/screens/core/constants/route_names.dart
@@ -1,1 +1,0 @@
-// TODO Implement this library.
