+++ lib/views/jobs/category_details.dart
